+++ nemo_skills/utils.py
@@ -350,3 +350,14 @@ def compute_chunk_ids(chunk_ids: list[int] | str, num_chunks: int) -> list[int] 
         assert chunk_id >= 0, "Run ids should have 1-based indexing"
 
     return chunk_ids
+
+
+def prefill_judgement(data_point: dict) -> str | None:
+    """Will automatically fill judgement if there is an exact match or the answer is None."""
+    if data_point['predicted_answer'] is None:
+        return "Reasoning: No answer was provided.\nJudgement: No"
+
+    if str(data_point['predicted_answer']).strip() == str(data_point['expected_answer']).strip():
+        return "Reasoning: The two answers are identical.\nJudgement: Yes"
+
+    return None
